added authentication to store deletion
